LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/coop-sandbox.https-expected.txt

CONSOLE MESSAGE: Navigation was blocked by Cross-Origin-Opener-Policy
CONSOLE MESSAGE: Navigation was blocked by Cross-Origin-Opener-Policy

PASS <iframe sandbox="allow-popups allow-scripts allow-same-origin"> Sandboxed Cross-Origin-Opener-Policy popup should result in a network error
PASS <iframe sandbox="allow-popups allow-scripts"> Sandboxed Cross-Origin-Opener-Policy popup should result in a network error
PASS <iframe sandbox="allow-popups allow-scripts allow-popups-to-escape-sandbox"> Sandboxed Cross-Origin-Opener-Policy popup should result in a network error
PASS Iframe with sandbox and COOP must load.
